+++ src/main/java/web/mvc/domain/QaBoardReply.java
@@ -14,7 +14,7 @@ public class QaBoardReply {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "reply_seq")
-    private Integer replySeq;
+    private Long replySeq;
 
     @Column(name = "comment", nullable = false, length = 255)
     private String comment;
